UI 4-15
Added some UI functionality to the game. The ability to click on a
button to control a worker unit is now in. It's a very crude first
iteration due to a lack of time. Hopefully this will be changed soon.
Players can also now click on different objects to see what they
are/contain.